@@ -10,19 +10,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.12",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.3",
     "@angular/cli": "~7.3.4",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/core": "^7.2.15",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "~7.2.0",
+    "@angular/forms": "~8.2.14",
     "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
     "@ng-idle/core": "^6.0.0-beta.4",
     "@ng-idle/keepalive": "^6.0.0-beta.4",
     "@ngu/carousel": "^1.5.5",
@@ -49,14 +49,14 @@
     "ngx-spinner": "^8.0.3",
     "ngx-toastr": "^10.0.2",
     "ngx-ui-loader": "^7.2.2",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.3",
     "tslib": "^1.9.0",
-    "typescript": "~3.2.2",
-    "zone.js": "~0.8.26"
+    "typescript": "~3.5.3",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular/language-service": "~8.2.14",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
